requirements.txt
fastapi>=0.100.0
fastapi[standard]>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
